requirement.txt
asgiref==3.2.3
astroid==2.3.3
backcall==0.1.0
decorator==4.4.1
Django==3.0.2
django-filter==1.1.0
djangorestframework==3.11.0
ipdb==0.12.3
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.2
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==3.3.0
pkg-resources==0.0.0
prompt-toolkit==3.0.2
ptyprocess==0.6.0
Pygments==2.5.2
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
traitlets==4.3.3
typed-ast==1.4.1
wcwidth==0.1.8
wrapt==1.11.2
